Implement client OutputWriter class (see details in description)
Follows an implementation similar to the previous lab, except that the send/receive channels are separated into classes that run on different threads. More try/catch statements are used than in the previous implementation to make errors codes more specific.
In addition, the socket code has been modified so it does not close but shuts down the output stream instead. The program no longer runs serially, so closing the socket in one thread could prevent data still leaving the server from entering the stream back to the client.